+++ commonjs/example/windows/status/table.js
@@ -1,6 +1,9 @@
 Ti.include(
     'create.js',
+    'update.js',
     'query.js',
+    'remove.js',
+    'search.js',
     'show.js'
 );
 
@@ -13,7 +16,7 @@ windowFunctions['Status'] = function (evt) {
         data: createRows([
             'Create Status',
             'Query Status',
-            'Show Status for User'
+            'Search Statuses by User'
         ])
     });
     table.addEventListener('click', handleOpenWindow);