+++ commonjs/documentation/changelog.md
@@ -4,6 +4,7 @@ v2.3.8  Updated the example app to use the Facebook module since it was pulled o
         Updated the example app lauout to work with iOS 7
         Added 3 new apis on push_schedules(create, query, delete) [CLOUDSRV-2785]
         Fixed example app issue entering ip address [TIMODOPEN-294]
+        Adding Users.resendConfirmation method [TIMODOPEN-293]
 
 v2.3.7  Added 3 new apis on push_notificiation(subsribe_token, unsubsribe_token, notify_tokens) [CLOUDSRV-2211][CLOUDSRV-2447]