+++ commonjs/documentation/changelog.md
@@ -5,6 +5,7 @@ v2.3.8  Updated the example app to use the Facebook module since it was pulled o
         Added 3 new apis on push_schedules(create, query, delete) [CLOUDSRV-2785]
         Fixed example app issue entering ip address [TIMODOPEN-294]
         Adding Users.resendConfirmation method [TIMODOPEN-293]
+        Added new apis on statuses(update, show, delete) [TIMODOPEN-296]
 
 v2.3.7  Added 3 new apis on push_notificiation(subsribe_token, unsubsribe_token, notify_tokens) [CLOUDSRV-2211][CLOUDSRV-2447]