+++ commonjs/documentation/changelog.md
@@ -9,6 +9,7 @@ v2.3.8  Updated the example app to use the Facebook module since it was pulled o
         Added Chats methods(remove, queryChatGroups) [TIMODOPEN-291]
         Added PushNotifications method (query) [TIMODOPEN-292]
         Added PushNotifications methods (queryChannels, showChannels) [TIMODOPEN-270]
+        Added likes apis(create, remove) [TIMODOPEN-289]
 
 v2.3.7  Added 3 new apis on push_notificiation(subsribe_token, unsubsribe_token, notify_tokens) [CLOUDSRV-2211][CLOUDSRV-2447]